pyproject.toml
[tool.poetry]
name = "test-task-hammer-systems"
version = "0.1.0"
description = "Simple referral api"
authors = ["Dmitry Ibragimov"]
readme = "README.md"
packages = [{include = "referral_app"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
djangorestframework-jwt = "^1.11.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.7"
gunicorn = "^21.2.0"
drf-yasg = {extras = ["validation"], version = "^1.21.7"}


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"

[tool.black]
line_lenght = 120

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
